arm/trusted-firmware-a: override do_compile to local source changes
trusted-firmware-a.inc[1] modifies the trusted-firmware-a source code
using sed in the do_compile function. This causes the source code to
be dirty and print the following:

    NOTICE:  BL31: v2.8(release):v2.8-226-g2fcd408bb3-dirty

As a result, it is not possible to know the original source code used
to build an image.

The local changes are only for building the native tools and the native
tools are not needed by VAR-SOM-AM62.

This patches removes the local changes to the source code and therefore
removes the -dirty suffix.

References:
[1] https://git.yoctoproject.org/meta-arm/tree/meta-arm/recipes-bsp/trusted-firmware-a/trusted-firmware-a.inc?h=dunfell#n103

Signed-off-by: Nate Drude <nate.d@variscite.com>