bcm43xx-utils: am62x-var-som: determine gpiochip using gpio label
The gpiochip, e.g. gpiochip1, is not stable because it changes when
gpio (e.g. i2c gpio expanders) are added or removed from the device tree.

This is problematic since it requires this script to be modified
when custom carrier boards add or remove i2c expanders.

Instead of hardcoding the gpiochip, use the gpio label, which is
determined by the kernel:
- For gpio controllers, the name is the address provided by the
  device tree. e.g. [600000.gpio]
- For i2c expanders, the name is the i2c bus and address, e.g. [0-0020]

Signed-off-by: Nate Drude <nate.d@variscite.com>