external-arm-toolchain: Introduce workaround for broken gconv packaging
external-arm-toolchain acts a lot like the typical glibc recipe found in
oe-core, however, it purportedly also provides components such as gconv,
normally found in glibc-locale. It relies directly on glibc bitbake code
to achieve various glibc functionality, but obtains its binaries from
pre-built artifacts installed on the host machine.

While the source may contain artifacts like gconvs, in reality, the
actual packaging of such artifacts is severely broken. This will
obviously lead to issues in a system that requires such components. For
reference, normal glibc systems package such artifacts utilizing the
libc-package.bbclass as follows:
https://github.com/openembedded/openembedded-core/blob/daaee6fcb0d2/meta/classes/libc-package.bbclass#L118-L138

To fix the broken packages, introduce a bbappend, which performs very
basic gconv packaging of the pre-built artifacts. This is not a true
fix, as doing so would require major re-work to the recipe at hand.
However, this should be sufficient for basic requirements of this system
and gconv usage.

Signed-off-by: Ken Sloat <ken.s@variscite.com>