external arm toolchain: fix libraries packaging
Due to a wrong sequence, some of the libraries are packed as part of
external-arm-toolchain-dev instead of the right package.
When building the SDK using populate-sdk, the following error is raised
nothing provides libatomic-dev needed by packagegroup-core-standalone-sdk-target
These changes are ported from meta-arm
https://git.yoctoproject.org/meta-arm/commit/?h=dunfell&id=f80387554cd9c706a25e031d16e61e4e44ba6d9c
They are specific for Dunfell branches of TI Processor SDKs and should
not be ported in the upcoming Kirkstone branches.

Signed-off-by: Pierluigi Passaro <pierluigi.p@variscite.com>